+++ runtime/elem/FloatBufferPool.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <memory>
 #include <stack>
 #include <unordered_map>
